+++ src/cascade/saver/db/save_model_results_to_db.py
@@ -2,6 +2,7 @@
 Saves results from a model fit to a database which the EpiViz can access.
 """
 import logging
+from pathlib import Path
 import tempfile
 
 try:
@@ -14,33 +15,10 @@ except ImportError:
 
     save_results_at = DummySaveResults()
 
-from cascade.saver.generate_draws import generate_draws_table
-
 DRAWS_INPUT_FILE_PATTERN = "all_draws.h5"
 
 CODELOG = logging.getLogger(__name__)
 MATHLOG = logging.getLogger(__name__)
-
-
-def save_model_results(execution_context):
-    """
-    We want to visualize model results using EpiViz.  To do this, we need to
-    generate draws from a model fit, and save those draws in a database
-    that EpiViz can access.
-
-    Args:
-        execution_context (ExecutionContext): contains model id data
-
-    """
-
-    dm_file = execution_context.dismodfile
-
-    if not dm_file:
-        raise ValueError("Must provide a dismod file")
-
-    draws_df = generate_draws_table(dm_file)
-
-    write_temp_draws_file_and_upload_model_results(draws_df, execution_context)
 
 
 def write_temp_draws_file_and_upload_model_results(draws_df, execution_context):
@@ -57,10 +35,10 @@ def write_temp_draws_file_and_upload_model_results(draws_df, execution_context):
 
     with tempfile.TemporaryDirectory() as tmpdirname:
 
-        file_path = tmpdirname + DRAWS_INPUT_FILE_PATTERN
+        file_path = (Path(tmpdirname) / DRAWS_INPUT_FILE_PATTERN)
 
         draws_df.to_hdf(
-            file_path,
+            str(file_path),
             "draws",
             format="table",
             data_columns=[
@@ -70,7 +48,7 @@ def write_temp_draws_file_and_upload_model_results(draws_df, execution_context):
                 "sex_id",
                 "year_id"])
 
-        CODELOG.debug("Saving Results")
+        CODELOG.debug("Saving Results to DB")
 
         modelable_entity_id = execution_context.parameters.modelable_entity_id
         model_title = execution_context.parameters.model_title
@@ -79,14 +57,14 @@ def write_temp_draws_file_and_upload_model_results(draws_df, execution_context):
         db_env = execution_context.parameters.db_env
 
         model_version_id_df = save_results_at(
-            str(tmpdirname),
+            tmpdirname,
             DRAWS_INPUT_FILE_PATTERN,
             modelable_entity_id,
             model_title,
             measures_to_save,
-            model_version_id,
+            model_version_id=model_version_id,
             db_env=db_env)
 
-        CODELOG.debug(f"mvid_df: {model_version_id_df.iloc[0, 0]}")
+        CODELOG.debug(f"model_version_id_df: {model_version_id_df.iloc[0, 0]}")
 
         return int(model_version_id_df.iloc[0, 0])
